+++ server/api/private/phrases/_phraseId@string/index.ts
@@ -1,0 +1,13 @@
+import type { DefineMethods } from 'aspida';
+import type { PhraseDto, PhraseUpdateVal } from 'common/types/phrase';
+
+export type Methods = DefineMethods<{
+  post: {
+    reqFormat: FormData;
+    reqBody: PhraseUpdateVal;
+    resBody: PhraseDto;
+  };
+  delete: {
+    resBody: PhraseDto;
+  };
+}>;
