+++ server/api/private/phrases/_phraseId@string/validators.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import { defineValidators } from './$relay';
+
+export default defineValidators(() => ({
+  params: z.object({
+    phraseId: z.string(),
+  }),
+}));
