--- server/api/private/phrases/_phrasesId@string/index.ts
@@ -1,17 +1,0 @@
-import type { DefineMethods } from 'aspida';
-import type { DtoId } from 'common/types/brandedId';
-import type { PhraseDto, PhraseUpdateDto } from 'common/types/phrase';
-
-export type Methods = DefineMethods<{
-  post: {
-    reqFormat: FormData;
-    reqBody: PhraseUpdateDto;
-    resBody: PhraseDto;
-  };
-  delete: {
-    reqBody: {
-      phraseId: DtoId['phrase'];
-    };
-    resBody: PhraseDto;
-  };
-}>;
